ecommerce

+++ ecommerce/src/main/java/com/grupo5/ecommerce/services/CategoriaService.java
@@ -3,44 +3,47 @@ package com.grupo5.ecommerce.services;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import com.grupo5.ecommerce.entities.Categoria;
 import com.grupo5.ecommerce.repositories.CategoriaRepository;
 
+@Service
 public class CategoriaService {
-
+	
 	@Autowired
 	CategoriaRepository categoriaRepository;
 	
 	public List<Categoria> getAllCategorias() {
-		return categoriaRepository.findAll();
+		List<Categoria> categorias = categoriaRepository.findAll();
+		return categorias;
 	}
 	
 	public Categoria getCategoriaById(Integer id) {
-		return categoriaRepository.findById(id).orElse(null);
+		Categoria categoria = categoriaRepository.findById(id).orElse(null);
+		
+		if(categoria==null)
+			return null;
+		
+		return categoria;
 	}
 	
 	public Categoria saveCategoria(Categoria categoria) {
 		return categoriaRepository.save(categoria);
 	}
 	
-	
-	public Categoria updateCategoria(Categoria categoria, Integer id) {
+	public Categoria updateCategoria(Categoria categoria) {
 		return categoriaRepository.save(categoria);
 	}
 	
-	public void deleteCategoria(Integer id) {
-		categoriaRepository.deleteById(id);
-	}
-	
-	public Boolean delCategoria(Integer id) {
-		categoriaRepository.deleteById(id);
-		Categoria categoriaDeletado = categoriaRepository.findById(id).orElse(null);
-		if (null == categoriaDeletado) {
-			return true;
-		}else {
-			return false;
-		}
-	}
+	   public Boolean delCategoria(Integer id) {
+		  if(categoriaRepository.findById(id).orElse(null)!=null) {
+			  categoriaRepository.deleteById(id);
+			  if(categoriaRepository.findById(id).orElse(null)==null)
+				  return true;
+		     else
+		    	 return false;
+		  }
+		    else return false;
+	      }
 }
-
