+++ ecommerce/src/main/java/com/grupo5/ecommerce/services/ProdutoService.java
@@ -3,44 +3,56 @@ package com.grupo5.ecommerce.services;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import com.grupo5.ecommerce.entities.Produto;
 import com.grupo5.ecommerce.repositories.ProdutoRepository;
 
+@Service
 public class ProdutoService {
-
+	
 	@Autowired
 	ProdutoRepository produtoRepository;
 	
 	public List<Produto> getAllProdutos() {
-		return produtoRepository.findAll();
+		List<Produto> produtos = produtoRepository.findAll();
+				
+		return produtos;
 	}
 	
 	public Produto getProdutoById(Integer id) {
-		return produtoRepository.findById(id).orElse(null);
+		Produto produto = produtoRepository.findById(id).orElse(null);
+		
+		if(produto==null)
+			return null;
+		
+		return produto;
+		
 	}
 	
 	public Produto saveProduto(Produto produto) {
 		return produtoRepository.save(produto);
+		 
 	}
 	
+	public Produto updateProduto(Produto produto) {
 	
-	public Produto updateProduto(Produto produto, Integer id) {
 		return produtoRepository.save(produto);
+		
+		
 	}
 	
-	public void deleteProduto(Integer id) {
-		produtoRepository.deleteById(id);
-	}
+	   public Boolean delProduto(Integer id) {
+		  if(produtoRepository.findById(id).orElse(null)!=null) {
+			  produtoRepository.deleteById(id);
+			  if(produtoRepository.findById(id).orElse(null)==null)
+				  return true;
+		     else
+		    	 return false;
+		  }
+		    else return false;
+	    	  
+	      }
 	
-	public Boolean delProduto(Integer id) {
-		produtoRepository.deleteById(id);
-		Produto produtoDeletado = produtoRepository.findById(id).orElse(null);
-		if (null == produtoDeletado) {
-			return true;
-		}else {
-			return false;
-		}
-	}
-}
 
+}
